% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count2fpkm.R
\name{count2fpkm}
\alias{count2fpkm}
\title{normalize count to FPKM}
\usage{
count2fpkm(count, geneLength, pseudoCount = 0, log = F, pseudoFPKM = 1)
}
\arguments{
\item{count}{a matrix or data.frame with unique id as row names and unique sample name as colnames required}

\item{geneLength}{a named vector with unique id (required)}

\item{pseudoCount}{the integer pseudoCount added to count matrix/data.frame. default = 0}

\item{log}{a boolean value to determine whether the output fpkm is Logarithm. default = FALSE. If is TRUE, use pseudoFPKM.}

\item{pseudoFPKM}{a integer pseudoCount added to FPKM matrix/data.frame after normalization for Logarithm. default = 1}
}
\value{
a data.frame of normalized count as FPKM
}
\description{
normalize count to FPKM
}
