% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count2tpm.R
\name{count2tpm}
\alias{count2tpm}
\title{normalize count to TPM}
\usage{
count2tpm(count, geneLength, pseudoCount = 0, log = F, pseudoTPM = 1)
}
\arguments{
\item{count}{a matrix or data.frame with unique id as row names and unique sample name as colnames required}

\item{geneLength}{a named vector with unique id (required)}

\item{pseudoCount}{the integer pseudoCount added to count matrix/data.frame. default = 0}

\item{log}{a boolean value to determine whether the output tpm is Logarithm. default = FALSE. If is TRUE, use pseudoTPM.}

\item{pseudoTPM}{a integer pseudoCount added to TPM matrix/data.frame after normalization for Logarithm. default = 1}
}
\value{
a data.frame of normalized count as TPM
}
\description{
normalize count to TPM
}
