% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plot_pca}
\alias{plot_pca}
\title{plot PCA with given normalized count matrix and sample annotation}
\usage{
plot_pca(
  tpm,
  col_anno,
  color_col,
  shape_col = NULL,
  text_col = NULL,
  point_size = 5,
  plot_file = ""
)
}
\arguments{
\item{tpm}{named count matrix/data.frame with column name representing sample names and row name representing gene name/id (required)}

\item{col_anno}{the column (sample) annotation data.frame with row name match the tpm matrix column name,
column name represents the factor to label for PCA plot (required)}

\item{color_col}{a column name from col_anno that be color aesthetics for PCA (required)}

\item{shape_col}{a column name from col_anno that be shape aesthetics for PCA. Default = NULL}

\item{text_col}{a column name from col_anno that be text label aesthetics for PCA. Default = NULL}

\item{point_size}{sample point size. Default = 5}

\item{plot_file}{save file to pdf. Default = ""}
}
\value{
a list of 4 plots, PCA matrix and PC variance summary. list(pc=pc, pca_summary=pca_summary, plots=plots)
}
\description{
plot PCA with given normalized count matrix and sample annotation
}
